Created a dedicated test for the Lines class
